reconstruction/point_cloud_processor.py
import numpy as np
from .utils.error_handling import InputProcessingError

class PointCloudProcessor:
    """
    Processes input data into a standardized point cloud format.

    This class handles various input types (file paths, numpy arrays, JSON-like structures)
    and converts them into a consistent point cloud representation for further processing.
    """

    def process(self, input_data):
        """
        Process the input data into a standardized point cloud format.

        Args:
            input_data: The input data, which can be a file path, numpy array, or JSON-like structure.

        Returns:
            numpy.ndarray: The processed point cloud data.

        Raises:
            InputProcessingError: If the input type is unsupported or processing fails.
        """
        try:
            if isinstance(input_data, str):
                return self.load_from_file(input_data)
            elif isinstance(input_data, np.ndarray):
                return self.process_numpy_array(input_data)
            elif isinstance(input_data, dict):
                return self.process_json_data(input_data)
            else:
                raise InputProcessingError("Unsupported input type")
        except Exception as e:
            raise InputProcessingError(f"Failed to process input data: {str(e)}")

    def load_from_file(self, file_path):
        """
        Load point cloud data from a file.

        Args:
            file_path (str): Path to the file containing point cloud data.

        Returns:
            numpy.ndarray: The loaded point cloud data.
        """
        # Implementation depends on the file format you're supporting
        # For example, for a CSV file:
        return np.loadtxt(file_path, delimiter=',')

    def process(self, input_data):
        """
        Process the input data into a standardized point cloud format.

        Args:
            input_data: The input data, which can be a file path (str), numpy array, list, or dictionary.

        Returns:
            numpy.ndarray: The processed point cloud data.

        Raises:
            InputProcessingError: If the input type is unsupported or processing fails.
        """
        try:
            if isinstance(input_data, str):
                return self.load_from_file(input_data)
            elif isinstance(input_data, np.ndarray):
                return self.process_numpy_array(input_data)
            elif isinstance(input_data, list):
                return self.process_numpy_array(np.array(input_data))
            elif isinstance(input_data, dict):
                return self.process_json_data(input_data)
            else:
                raise InputProcessingError("Unsupported input type")
        except Exception as e:
            raise InputProcessingError(f"Failed to process input data: {str(e)}")

    def process_numpy_array(self, array):
        """
        Process a numpy array containing point cloud data.

        Args:
            array (numpy.ndarray): The input numpy array.

        Returns:
            numpy.ndarray: The processed point cloud data.

        Raises:
            InputProcessingError: If the array shape is invalid.
        """
        # Ensure the array has the correct shape and type
        if array.ndim != 2 or array.shape[1] not in [3, 6]:
            raise InputProcessingError("Invalid array shape. Expected Nx3 or Nx6.")
        return array.astype(np.float32)

    def process_json_data(self, data):
        """
        Process JSON-like data containing point cloud information.

        Args:
            data (dict): A dictionary containing point cloud data.

        Returns:
            numpy.ndarray: The processed point cloud data.
        """
        # Example implementation - adjust based on your JSON structure
        if 'points' not in data:
            raise InputProcessingError("JSON data must contain 'points' key")
        return np.array(data['points'], dtype=np.float32)
